Deleted expert_QC.pynb, moved quality visualisation to functions.py
