+++ envass/functions.py
@@ -50,3 +50,50 @@ def interp_nan(time, y):
     nans, x = np.isnan(vec), lambda z: z.nonzero()[0]
     vec[nans]=np.interp(time[nans], time[~nans],vec[~nans])
     return vec
+
+def visualize_quality_assurance(df_sub):
+    import plotly.graph_objs as go
+    import plotly.offline as py
+    from ipywidgets import interactive, HBox, VBox
+
+    variables=df_sub.columns
+
+
+    py.init_notebook_mode()
+    f = go.FigureWidget([go.Scatter(y = df_sub.index, x = df_sub.index, mode = 'markers')])
+    scatter = f.data[0]
+    N = len(df_sub)
+    scatter.marker.opacity = 0.8
+    def update_axes(xaxis, yaxis):
+        scatter = f.data[0]
+        scatter.x = df_sub[xaxis]
+        scatter.y = df_sub[yaxis]
+
+        with f.batch_update():
+            f.layout.xaxis.title = xaxis
+            f.layout.yaxis.title = yaxis
+            if "_qual" not in yaxis:
+                f.add_trace(go.Scatter(y = df_sub[yaxis][df_sub[yaxis+"_qual"]==0], x = df_sub[xaxis][df_sub[yaxis+"_qual"]==0], mode = 'markers', marker = dict(color = 'blue'), name = f'{yaxis} Trusted (=0)'))
+                f.add_trace(go.Scatter(y = df_sub[yaxis][df_sub[yaxis+"_qual"]==1], x = df_sub[xaxis][df_sub[yaxis+"_qual"]==1], mode = 'markers', marker = dict(color = 'darkred'), name = f'{yaxis} Not trusted (=1)'))
+            else: 
+                scatter.x = df_sub[xaxis]
+                scatter.y = df_sub[yaxis]
+
+    axis_dropdowns = interactive(update_axes, yaxis = df_sub.columns, xaxis = df_sub.columns)
+
+    # Create a table FigureWidget that updates on selection from points in the scatter plot of f
+    t = go.FigureWidget([go.Table(
+    header=dict(values=variables,
+                fill = dict(color='#C2D4FF'),
+                align = ['left'] * 5),
+    cells=dict(values=[df_sub[col] for col in variables],
+               fill = dict(color='#F5F8FF'),
+               align = ['left'] * 5))])
+
+    def selection_fn(trace,points,selector):
+        t.data[0].cells.values = [df_sub.loc[points.point_inds][col] for col in variables]
+
+    scatter.on_selection(selection_fn)
+
+    # Put everything together
+    return VBox((HBox(axis_dropdowns.children),f,t))
